significantly updated the website
I updated index.html to give more information on how to enable JavaScript. I also added a 404 page in case the website user goes to a page that doesn't exist. I also updated the website style to be more modern and the color scheme to be more dark than light mode. Last but not least, I added a GitHub button if the website user would like to view the source code of the website (I'm not hiding anything)